{feature} introduce background option
With this option, a background process is started with
a watching `ng build`. This is vastly more performant
than running a new `ng build` every time a change is made,
with build times in the tens of milliseconds instead of
a couple of seconds on my machine for the test-book.